+++ packages/emails/src/templates.ts
@@ -1,8 +1,8 @@
-export * from "./templates/new-comment";
 export * from "./previews/finalized-host";
 export * from "./previews/finalized-participant";
-export * from "./previews/login";
 export * from "./previews/new-participant";
 export * from "./previews/new-participant-confirmation";
 export * from "./previews/new-poll";
 export * from "./previews/register";
+export * from "./templates/login";
+export * from "./templates/new-comment";
