+++ packages/emails/src/templates/new-participant-confirmation.tsx
