+++ packages/emails/src/templates/new-participant.tsx
