+++ packages/emails/src/templates/new-poll.tsx
